+++ src/popup.html
@@ -87,6 +87,6 @@
         <ul class="menu-list" id="history_list"></ul>
       </div>
     </div>
-    <script src="js/popup.js"></script>
+    <script type="module" src="js/popup.js"></script>
   </body>
 </html>
